requirements.txt
openai>=1.12.0
faiss-cpu>=1.7.4
langchain>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.35.0
playwright>=1.42.0
